pyproject.toml
[tool.poetry]
name = "notes"
version = "0.1.0"
description = ""
authors = ["Z1pP <maxm_99@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12"
fastapi = "^0.114.2"
sqlalchemy = "^2.0.34"
asyncpg = "^0.29.0"
black = "^24.8.0"
aiogram = "^3.13.0"
uvicorn = "^0.30.6"
pyjwt = {extras = ["crypto"], version = "^2.9.0"}
alembic = "^1.13.2"
pydantic-settings = "^2.5.2"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
